fix: offload FastMCP tools and publish verified badges [event-loop-badges-applied]
